duplicate code removal ( create subscription part)
