+++ src/js/modules/orders/orders.jsx
@@ -13,7 +13,7 @@ import Dropdown from 'src/js/components/dropdown';
 import FormTemplate from 'src/js/components/form/formTemplate';
 import OrdersNPositionsTab from 'src/js/components/ordersNPositionsTab';
 import { checkIfOption, checkIfPutCallExpiry } from 'src/js/utils/global';
-import { unSubscribe, fetchInfo } from 'src/js/utils/queries';
+import { unSubscribe, fetchInfo, subscribePrices } from 'src/js/utils/queries';
 
 class Orders extends React.PureComponent {
     constructor() {
@@ -106,7 +106,7 @@ class Orders extends React.PureComponent {
     }
 
     handleSubscribe(instrument) {
-        queries.createSubscription(instrument, this.props, this.handlePriceUpdate, (subscription) => {
+        subscribePrices(this.props, instrument, this.handlePriceUpdate, (subscription) => {
             this.subscription = subscription;
         });
     }
