+++ src/js/modules/tradeSubscription/queries.js
@@ -1,14 +1,4 @@
-import { doWithLoader } from 'src/js/utils/global';
-import * as API from 'src/js/utils/api';
 import _ from 'lodash';
-
-export function createSubscription(props, subscriptionArgs, tradeType, onUpdate, cb) {
-    doWithLoader(
-        props,
-        _.partial(API[`create${tradeType}Subscription`], props.accessToken, subscriptionArgs, onUpdate),
-        (result) => cb(result)
-    );
-}
 
 export function getUpdatedTrades(currTrades, tradeTypeId, updatedTrades) {
     _.forEach(updatedTrades, (value, index) => {
